# Conflicts:
#	pom.xml
#	smf-capi/src/main/java/com/smf/main/ApplicationStarter.java
#	smf-domain/src/main/resources/application.properties

+++ smf-model/src/main/java/com/smf/main/model/FundRegistration.java
@@ -12,5 +12,5 @@ import lombok.*;
 @AllArgsConstructor
 public class FundRegistration {
     private String fundName;
-    private Long fundAmount;
+    private Double fundAmount;
 }
